requirements.txt
Flask==3.0.3
Flask-Cors==5.0.1
python-dotenv==1.1.0
openai==1.30.1
pdfplumber==0.11.6
gunicorn==21.2.0